user authentication JWT

+++ backend/controllers/groceryController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Grocery = require('../models/groceryModel')
+const User = require('../models/userModel')
 
 // @desc Get Groceries
 // @route GET /api/groceries
 // @access Private
 const getGroceries = asyncHandler(async (req, res) => {
-    const groceries = await Grocery.find()
+    const groceries = await Grocery.find({ user: req.user.id})
     res.status(200).json(groceries)
 })
 
@@ -21,6 +22,7 @@ const createGrocery = asyncHandler(async (req, res) => {
 
     const grocery  = await Grocery.create({
         name: req.body.name,
+        user: req.user.id
     })
 
     res.status(200).json(grocery)
@@ -35,6 +37,20 @@ const updateGrocery = asyncHandler(async (req, res) => {
     if(!grocery){
         res.status(400)
         throw new Error('Grocery item not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    // make sure the logged in user matches grocery user
+    if(grocery.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('user not authorized')
     }
 
     const updatedGrocery = 
@@ -56,12 +72,25 @@ const deleteGrocery = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Grocery item not found')
     }
+    
+    const user = await User.findById(req.user.id)
+
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error('user not found')
+    }
+
+    // make sure the logged in user matches grocery user
+    if(grocery.user.toString() !== user.id){
+        res.status(401)
+        throw new Error(`user not authorized`)
+    }
 
     await grocery.deleteOne()
 
     res.status(200).json({message: `Delete grocery id: ${req.params.id}`})
 })
-
 
 module.exports = {
     getGroceries, createGrocery, updateGrocery, deleteGrocery
